+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Fixed
 - Command Login - Disable account creation.
 - Command Login - Check token validity before login.
+- Command Generate - Prompt for password/token just after the other prompts.
 
 ## RELEASE 2.6.3 - 2019-11-04
 ### Fixed